Automatically expand parameters for analysis #772
