+++ src/PSRule.Rules.Azure/Pipeline/LoggingExtensions.cs
@@ -12,26 +12,41 @@ namespace PSRule.Rules.Azure.Pipeline
     {
         internal static void VerboseFindFiles(this ILogger logger, string path)
         {
+            if (logger == null)
+                return;
+
             logger.WriteVerbose(Diagnostics.VerboseFindFiles, path);
         }
 
         internal static void VerboseFoundFile(this ILogger logger, string path)
         {
+            if (logger == null)
+                return;
+
             logger.WriteVerbose(Diagnostics.VerboseFoundFile, path);
         }
 
         internal static void VerboseMetadataNotFound(this ILogger logger, string path)
         {
+            if (logger == null)
+                return;
+
             logger.WriteVerbose(Diagnostics.VerboseMetadataNotFound, path);
         }
 
         internal static void VerboseTemplateLinkNotFound(this ILogger logger, string path)
         {
+            if (logger == null)
+                return;
+
             logger.WriteVerbose(Diagnostics.VerboseTemplateLinkNotFound, path);
         }
 
         internal static void VerboseTemplateFileNotFound(this ILogger logger, string path)
         {
+            if (logger == null)
+                return;
+
             logger.WriteVerbose(Diagnostics.VerboseTemplateFileNotFound, path);
         }
     }
