+++ src/PSRule.Rules.Azure/Runtime/Helper.cs
@@ -1,7 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using PSRule.Rules.Azure.Configuration;
 using PSRule.Rules.Azure.Data.Template;
+using PSRule.Rules.Azure.Pipeline;
+using System.Management.Automation;
 
 namespace PSRule.Rules.Azure.Runtime
 {
@@ -21,6 +24,26 @@ namespace PSRule.Rules.Azure.Runtime
         public static bool IsTemplateExpression(string expression)
         {
             return ExpressionParser.IsExpression(expression);
+        }
+
+        public static PSObject[] GetResources(string parameterFile)
+        {
+            var context = GetContext();
+
+            var linkHelper = new TemplateLinkHelper(context, PSRuleOption.GetWorkingPath(), true);
+            var link = linkHelper.ProcessParameterFile(parameterFile);
+            if (link == null)
+                return null;
+
+            var helper = new TemplateHelper(context, "helper", context.Option.Configuration.ResourceGroup, context.Option.Configuration.Subscription);
+            return helper.ProcessTemplate(link.TemplateFile, link.ParameterFile);
+        }
+
+        private static PipelineContext GetContext()
+        {
+            var option = PSRuleOption.FromFileOrDefault(PSRuleOption.GetWorkingPath());
+            var context = new PipelineContext(option, null);
+            return context;
         }
     }
 }
