+++ tests/PSRule.Rules.Azure.Tests/Azure.Convention.Tests.ps1
@@ -1,0 +1,43 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+#
+# Unit tests for Azure conventions
+#
+
+[CmdletBinding()]
+param ()
+
+# Setup error handling
+$ErrorActionPreference = 'Stop';
+Set-StrictMode -Version latest;
+
+if ($Env:SYSTEM_DEBUG -eq 'true') {
+    $VerbosePreference = 'Continue';
+}
+
+# Setup tests paths
+$rootPath = $PWD;
+Import-Module (Join-Path -Path $rootPath -ChildPath out/modules/PSRule.Rules.Azure) -Force;
+$here = (Resolve-Path $PSScriptRoot).Path;
+
+Describe 'Azure.ExpandTemplate' -Tag 'Convention' {
+    Context 'Convention' {
+        It 'Expands template parameter files' {
+            $invokeParams = @{
+                Baseline = 'Azure.All'
+                Module = 'PSRule.Rules.Azure'
+                WarningAction = 'Ignore'
+                ErrorAction = 'Stop'
+            }
+
+            $parameterFile = Join-Path -Path $here -ChildPath 'Resources.Storage.Parameters.json';
+            $result = @(Invoke-PSRule @invokeParams -InputPath $parameterFile -Format File);
+            $result.Length | Should -BeGreaterThan 1;
+
+            $resource = $result | Where-Object { $_.TargetType -eq 'Microsoft.Storage/storageAccounts' };
+            $resource | Should -Not -BeNullOrEmpty;
+            $resource.TargetName | Should -BeIn 'storage1'
+        }
+    }
+}
